twilio-sms-manager-mvp/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.9

SQLAlchemy==2.0.36
alembic==1.13.3

passlib[bcrypt]==1.7.4
bcrypt<4
python-jose[cryptography]==3.3.0

twilio==9.4.4
httpx==0.27.2

streamlit==1.41.1
streamlit-autorefresh==1.0.1
pandas==2.2.3
python-dateutil==2.9.0.post0
